requirements.txt
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
psutil==5.9.5
pyasn1==0.5.0
pydantic==1.10.7
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==4.5.4
redislite==6.2.805324
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
starlette==0.26.1
typing-extensions==4.5.0
uvicorn==0.21.1
